Instantiate sort_by_key kernels in separately
