+++ src/backend/cuda/CMakeLists.txt
@@ -158,7 +158,7 @@ FILE(GLOB cuda_headers
 FILE(GLOB cuda_sources
     "*.cu"
     "*.cpp"
-    "sort_by_key/*.cu"
+    "kernel/sort_by_key/*.cu"
     "kernel/*.cu")
 
 FILE(GLOB jit_sources
